correcciones

+++ src/main/java/com/example/sintomedic/exception/SintomaNotFoundException.java
@@ -1,6 +1,6 @@
 package com.example.sintomedic.exception;
 
 public class SintomaNotFoundException extends Exception {
-    public SintomaNotFoundException(Long id) {
+    public SintomaNotFoundException(int id) {
     }
 }
