+++ src/main/java/com/example/sintomedic/repositorios/SintomasRepositorio.java
@@ -4,7 +4,7 @@ import com.example.sintomedic.Sintoma;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface SintomasRepositorio extends JpaRepository<Sintoma,Long> {
+public interface SintomasRepositorio extends JpaRepository<Sintoma,Integer> {
 
 
 }
